// shared/pagination/pagination-sort.enum.ts
export const enum sortBy {
  NameAsc = 'name-asc',
  NameDesc = 'name-desc',
  SecretDateDesc = 'secret-date-desc',
  SecretDateAsc = 'secret-date-asc',
  KeyDateDesc = 'key-date-desc',
  KeyDateAsc = 'key-date-asc',
  ActiveDesc = 'active-desc',
  ActiveAsc = 'active-asc',
  OrderAsc = 'order-asc',
  OrderDesc = 'order-desc',
}